# backend/offertracker-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OfferTracker API",
    "description": "求职管理系统后端API",
    "version": "1.0.0",
    "contact": {
      "name": "OfferTracker Team",
      "url": "https://github.com/offertracker"
    }
  },
  "servers": [
    {
      "url": "https://offertracker-backend.1544718463.workers.dev",
      "description": "生产环境"
    },
    {
      "url": "http://localhost:8787",
      "description": "开发环境"
    }
  ],
...
  "paths": {
...
    "/api/register": {
      "post": {
        "summary": "用户注册",
        "description": "创建新用户账户",
        "tags": ["认证"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "confirmPassword"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "password123"
                  },
                  "confirmPassword": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "summary": "用户登录",
        "description": "用户登录获取访问令牌",
        "tags": ["认证"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "认证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "summary": "获取当前用户信息",
        "description": "获取当前登录用户的信息",
        "tags": ["认证"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "成功获取用户信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "未授权访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/applications": {
      "get": {
        "summary": "获取求职记录列表",
        "description": "获取当前用户的所有求职记录",
        "tags": ["求职记录"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "成功获取求职记录列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Application"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建求职记录",
        "description": "创建新的求职申请记录",
        "tags": ["求职记录"],
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Application"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/{id}": {
      "get": {
        "summary": "获取单条求职记录",
        "description": "获取指定的求职记录详情",
        "tags": ["求职记录"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取求职记录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Application"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "更新求职记录",
        "description": "更新指定的求职记录",
        "tags": ["求职记录"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Application"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除求职记录",
        "description": "删除指定的求职记录",
        "tags": ["求职记录"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "错误信息"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "jwt_token"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "user-id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "plan_type": {
            "type": "string",
            "enum": ["free", "pro", "team"],
            "example": "free"
          },
          "created_at": {
            "type": "integer",
            "example": 1234567890
          }
        }
      },
...
      "Application": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "application-id"
          },
          "user_id": {
            "type": "string",
            "example": "user-id"
          },
          "resume_id": {
            "type": "string",
            "example": "resume-id"
          },
          "company_name": {
            "type": "string",
            "example": "公司名称"
          },
          "position_title": {
            "type": "string",
            "example": "职位名称"
          },
          "status": {
            "type": "string",
            "example": "applied"
          },
          "city": {
            "type": "string",
            "example": "北京"
          },
          "salary_range": {
            "type": "string",
            "example": "20k-30k"
          },
          "channel": {
            "type": "string",
            "example": "boss"
          },
          "contact_name": {
            "type": "string",
            "example": "联系人"
          },
          "contact_email": {
            "type": "string",
            "format": "email",
            "example": "contact@example.com"
          },
          "application_date": {
            "type": "integer",
            "example": 1234567890
          },
          "last_update": {
            "type": "integer",
            "example": 1234567890
          },
          "interview_date": {
            "type": "integer",
            "example": 1234567890
          },
          "offer_status": {
            "type": "string",
            "example": "accepted"
          },
          "rejection_reason": {
            "type": "string",
            "example": "拒绝原因"
          },
          "notes": {
            "type": "string",
            "example": "备注信息"
          },
          "custom_fields": {
            "type": "string",
            "example": "{}"
          },
          "resume_name": {
            "type": "string",
            "example": "简历名称"
          }
        }
      },
      "ApplicationInput": {
        "type": "object",
        "properties": {
          "resume_id": {
            "type": "string",
            "example": "resume-id"
          },
          "company_name": {
            "type": "string",
            "example": "公司名称"
          },
          "position_title": {
            "type": "string",
            "example": "职位名称"
          },
          "status": {
            "type": "string",
            "example": "applied"
          },
          "city": {
            "type": "string",
            "example": "北京"
          },
          "salary_range": {
            "type": "string",
            "example": "20k-30k"
          },
          "channel": {
            "type": "string",
            "example": "boss"
          },
          "contact_name": {
            "type": "string",
            "example": "联系人"
          },
          "contact_email": {
            "type": "string",
            "format": "email",
            "example": "contact@example.com"
          },
          "application_date": {
            "type": "integer",
            "example": 1234567890
          },
          "interview_date": {
            "type": "integer",
            "example": 1234567890
          },
          "offer_status": {
            "type": "string",
            "example": "pending"
          },
          "rejection_reason": {
            "type": "string",
            "example": "拒绝原因"
          },
          "notes": {
            "type": "string",
            "example": "备注信息"
          },
          "custom_fields": {
            "type": "string",
            "example": "{}"
          }
        }
      },
...
    }
  }
}